WIP move net.imglib2.ui classes
InteractiveDisplayCanvas -> bdv.viewer
OverlayRenderer          -> bdv.viewer
