+++ src/main/java/bdv/viewer/ViewerPanel.java
@@ -82,8 +82,6 @@ import net.imglib2.RealLocalizable;
 import net.imglib2.RealPoint;
 import net.imglib2.RealPositionable;
 import net.imglib2.realtransform.AffineTransform3D;
-import net.imglib2.ui.InteractiveDisplayCanvas;
-import net.imglib2.ui.OverlayRenderer;
 import net.imglib2.ui.PainterThread;
 import net.imglib2.util.LinAlgHelpers;
 import org.scijava.listeners.Listeners;
