+++ src/main/java/net/imglib2/ui/overlay/BufferedImageOverlayRenderer.java
@@ -41,7 +41,7 @@ import java.awt.Graphics2D;
 import java.awt.RenderingHints;
 import java.awt.image.BufferedImage;
 import net.imglib2.realtransform.AffineTransform3D;
-import net.imglib2.ui.OverlayRenderer;
+import bdv.viewer.OverlayRenderer;
 import bdv.viewer.render.RenderTarget;
 import bdv.viewer.TransformListener;
 import org.scijava.listeners.Listeners;
